Results now show latest post at top. Delete All results now has a prompt to confirm delete

+++ src/components/Result/DeleteAllResults/DeleteAllResults.js
@@ -1,12 +1,13 @@
 import React, { Component } from 'react';
 import Aux from '../../../hoc/Auxiliary/Auxiliary';
 import axios from '../../../axios-orders';
-import classes from './DeleteAllResults.css'
-
+import classes from './DeleteAllResults.css';
 
 export default class ResultDeleteAll extends Component {
-
   deleteAllResultsHandler = () => {
+    if (window.confirm('Are you sure you wish to delete all your results?') === false) {
+      return;
+    }
     axios.delete('/assets.json').then(x => {
       window.location.reload();
     });
